// app/src/types/Inmates.ts
type Inmate = {
    FirstName: string;
    MiddleName: string;
    LastName: string;
    CIDNo: string;
    DOB: Date;
    Sex: string;
    Village: string;
    Gewog: string;
    Dzongkhag: string;
    MaritalStatus: string;
    Nationality: string;
    Height: number;
    Weight: number;
    ColorOfHair: string;
    ColorOfEyes: string;
    Complexion: string;
    CrimesCommitted: string;
    Sentence: string;
    TimeServeStarts: Date;
    TimeServeEnds: Date;
    PrisonId: number;
    Image: string;
    EmergencyPersonName: string;
    EmergencyPersonContact: string;
    EmergencyPersonRelationship: string;
  };

  export default Inmate;